youtube_video_downloader/requirements.txt
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
numpy==1.19.1
Pillow==7.2.0
proglog==0.1.9
pytube3==9.6.4
requests==2.24.0
tqdm==4.48.2
typing-extensions==3.7.4.2
UNKNOWN==0.0.0
urllib3==1.25.10
